+++ service/controller/resource/alert/resource.go
@@ -1,6 +1,7 @@
 package alert
 
 import (
+	"bytes"
 	"reflect"
 
 	"github.com/giantswarm/microerror"
@@ -8,14 +9,16 @@ import (
 	promv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	promclient "github.com/prometheus-operator/prometheus-operator/pkg/client/versioned"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"sigs.k8s.io/yaml"
 
-	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/alert/rules"
-	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/generic"
+	"github.com/giantswarm/prometheus-meta-operator/pkg/project"
+	"github.com/giantswarm/prometheus-meta-operator/pkg/template"
 	"github.com/giantswarm/prometheus-meta-operator/service/key"
 )
 
 const (
-	Name = "alert"
+	Name              = "alert"
+	rulesFileLocation = "/opt/prometheus-meta-operator/files/templates/rules/**/*.yml"
 )
 
 type Config struct {
@@ -24,7 +27,13 @@ type Config struct {
 	Logger           micrologger.Logger
 }
 
-func New(config Config) (*generic.Resource, error) {
+type Resource struct {
+	prometheusClient promclient.Interface
+	logger           micrologger.Logger
+	installation     string
+}
+
+func New(config Config) (*Resource, error) {
 	if config.PrometheusClient == nil {
 		return nil, microerror.Maskf(invalidConfigError, "%T.PrometheusClient must not be empty", config)
 	}
@@ -35,66 +44,61 @@ func New(config Config) (*generic.Resource, error) {
 		return nil, microerror.Maskf(invalidConfigError, "%T.Installation must not be empty", config)
 	}
 
-	clientFunc := func(namespace string) generic.Interface {
-		c := config.PrometheusClient.MonitoringV1().PrometheusRules(namespace)
-		return wrappedClient{client: c}
-	}
-
-	c := generic.Config{
-		ClientFunc:    clientFunc,
-		Logger:        config.Logger,
-		Name:          Name,
-		GetObjectMeta: getObjectMeta,
-		GetDesiredObject: func(obj interface{}) (metav1.Object, error) {
-			return getRules(obj, config.Installation)
-		},
-		HasChangedFunc: hasChanged,
-	}
-	r, err := generic.New(c)
-	if err != nil {
-		return nil, microerror.Mask(err)
+	r := &Resource{
+		logger:           config.Logger,
+		prometheusClient: config.PrometheusClient,
+		installation:     config.Installation,
 	}
 
 	return r, nil
 }
 
-func getObjectMeta(obj interface{}) (metav1.ObjectMeta, error) {
-	cluster, err := key.ToCluster(obj)
-	if err != nil {
-		return metav1.ObjectMeta{}, microerror.Mask(err)
-	}
-
-	return metav1.ObjectMeta{
-		Name:      "prometheus-meta-operator",
-		Namespace: key.Namespace(cluster),
-		Labels: map[string]string{
-			key.ClusterIDKey(): key.ClusterID(cluster),
-		},
-	}, nil
+func (r *Resource) Name() string {
+	return "alert"
 }
 
-func getRules(obj interface{}, installation string) (metav1.Object, error) {
-	objectMeta, err := getObjectMeta(obj)
-	if err != nil {
-		return nil, microerror.Mask(err)
-	}
+type TemplateData struct {
+	ClusterID    string
+	Installation string
+	ManagedBy    string
+	Namespace    string
+}
 
+func getRules(obj interface{}, installation string) ([]*promv1.PrometheusRule, error) {
 	cluster, err := key.ToCluster(obj)
 	if err != nil {
 		return nil, microerror.Mask(err)
 	}
 
-	r := &promv1.PrometheusRule{
-		ObjectMeta: objectMeta,
-		Spec: promv1.PrometheusRuleSpec{
-			Groups: []promv1.RuleGroup{
-				rules.LabellingSchemaValidationRule(cluster, installation),
-				rules.Heartbeat(cluster, installation),
-			},
-		},
+	var data TemplateData = TemplateData{
+		ClusterID:    key.ClusterID(cluster),
+		Installation: installation,
+		ManagedBy:    project.Name(),
+		Namespace:    key.Namespace(cluster),
 	}
 
-	return r, nil
+	template, err := template.RenderTemplate(data, rulesFileLocation)
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
+
+	// Split multi yaml files
+	files := bytes.Split(template, []byte("---"))
+
+	var rules []*promv1.PrometheusRule = make([]*promv1.PrometheusRule, 0)
+	for _, file := range files {
+		if len(bytes.TrimSpace(file)) > 0 {
+			var rule promv1.PrometheusRule = promv1.PrometheusRule{}
+
+			if err = yaml.Unmarshal(file, &rule); err != nil {
+				return nil, microerror.Mask(err)
+			}
+
+			rules = append(rules, &rule)
+		}
+	}
+
+	return rules, nil
 }
 
 func hasChanged(current, desired metav1.Object) bool {
@@ -102,4 +106,21 @@ func hasChanged(current, desired metav1.Object) bool {
 	d := desired.(*promv1.PrometheusRule)
 
 	return !reflect.DeepEqual(c.Spec, d.Spec)
+}
+
+func updateMeta(c, d metav1.Object) {
+	d.SetGenerateName(c.GetGenerateName())
+	d.SetUID(c.GetUID())
+	d.SetResourceVersion(c.GetResourceVersion())
+	d.SetGeneration(c.GetGeneration())
+	d.SetSelfLink(c.GetSelfLink())
+	d.SetCreationTimestamp(c.GetCreationTimestamp())
+	d.SetDeletionTimestamp(c.GetDeletionTimestamp())
+	d.SetDeletionGracePeriodSeconds(c.GetDeletionGracePeriodSeconds())
+	d.SetLabels(c.GetLabels())
+	d.SetAnnotations(c.GetAnnotations())
+	d.SetFinalizers(c.GetFinalizers())
+	d.SetOwnerReferences(c.GetOwnerReferences())
+	d.SetClusterName(c.GetClusterName())
+	d.SetManagedFields(c.GetManagedFields())
 }