+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix kube-state-metrics scraping port on Control Planes.
 
+### Added
+
+- Added recording rules
+
 ## [1.8.0] - 2020-10-21
 
 ### Added